+++ tests/core/library_test.py
@@ -87,6 +87,21 @@ class CoreLibraryTest(unittest.TestCase):
         self.library1.find_exact.assert_called_once_with(any=['a'])
         self.library2.find_exact.assert_called_once_with(any=['a'])
 
+    def test_find_accepts_query_dict_instead_of_kwargs(self):
+        track1 = Track(uri='dummy1:a')
+        track2 = Track(uri='dummy2:a')
+        self.library1.find_exact().get.return_value = [track1]
+        self.library1.find_exact.reset_mock()
+        self.library2.find_exact().get.return_value = [track2]
+        self.library2.find_exact.reset_mock()
+
+        result = self.core.library.find_exact(dict(any=['a']))
+
+        self.assertIn(track1, result)
+        self.assertIn(track2, result)
+        self.library1.find_exact.assert_called_once_with(any=['a'])
+        self.library2.find_exact.assert_called_once_with(any=['a'])
+
     def test_search_combines_results_from_all_backends(self):
         track1 = Track(uri='dummy1:a')
         track2 = Track(uri='dummy2:a')
@@ -96,6 +111,21 @@ class CoreLibraryTest(unittest.TestCase):
         self.library2.search.reset_mock()
 
         result = self.core.library.search(any=['a'])
+
+        self.assertIn(track1, result)
+        self.assertIn(track2, result)
+        self.library1.search.assert_called_once_with(any=['a'])
+        self.library2.search.assert_called_once_with(any=['a'])
+
+    def test_search_accepts_query_dict_instead_of_kwargs(self):
+        track1 = Track(uri='dummy1:a')
+        track2 = Track(uri='dummy2:a')
+        self.library1.search().get.return_value = [track1]
+        self.library1.search.reset_mock()
+        self.library2.search().get.return_value = [track2]
+        self.library2.search.reset_mock()
+
+        result = self.core.library.search(dict(any=['a']))
 
         self.assertIn(track1, result)
         self.assertIn(track2, result)